desugar_match.rs hands the grouped variant match to desugar_match_grouped.rs, a pure text move under the 800-line ceiling
